Change FE section to be about static FE dev
This year we had a stronger separation between HTML/CSS and JS.
The goal of this change is to create a firmer sense of role for each
tech. It also plays nicely into layering the FE curriculum.
We start off building static, brochure sites and graduate to
progressively enhanced websites with JS.
Maybe at some point we can include something like React in the
curriculum to show a more complete evolution of the FE space.